Fix warning: implicit conversion changes signedness
In file included from visp/modules/tracker/rbt/src/features/vpRBDenseDepthTracker.cpp:31:
visp/modules/tracker/rbt/include/visp3/rbt/vpRBDenseDepthTracker.h: In member function ‘void vpRBDenseDepthTracker::vpDepthPointSet::updateAndErrorAndInteractionMatrix(const vpCameraParameters&, const vpHomogeneousMatrix&, const vpImage<float>&, vpColVector&, vpMatrix&)’:
visp/modules/tracker/rbt/include/visp3/rbt/vpRBDenseDepthTracker.h:223:21: warning: conversion to ‘std::vector<bool>::size_type’ {aka ‘long unsigned int’} from ‘int’ may change the sign of the result [-Wsign-conversion]
  223 |             m_valid[i] = false;
      |                     ^
visp/modules/tracker/rbt/include/visp3/rbt/vpRBDenseDepthTracker.h:233:21: warning: conversion to ‘std::vector<bool>::size_type’ {aka ‘long unsigned int’} from ‘int’ may change the sign of the result [-Wsign-conversion]
  233 |             m_valid[i] = false;
      |                     ^
visp/modules/tracker/rbt/include/visp3/rbt/vpRBDenseDepthTracker.h:239:21: warning: conversion to ‘std::vector<bool>::size_type’ {aka ‘long unsigned int’} from ‘int’ may change the sign of the result [-Wsign-conversion]
  239 |             m_valid[i] = false;
      |                     ^
visp/modules/tracker/rbt/include/visp3/rbt/vpRBDenseDepthTracker.h:243:19: warning: conversion to ‘std::vector<bool>::size_type’ {aka ‘long unsigned int’} from ‘int’ may change the sign of the result [-Wsign-conversion]
  243 |           m_valid[i] = true;
      |                   ^
visp/modules/tracker/rbt/include/visp3/rbt/vpRBDenseDepthTracker.h:256:13: warning: conversion to ‘unsigned int’ from ‘int’ may change the sign of the result [-Wsign-conversion]
  256 |           e[i] = D + projNormal;
      |             ^
visp/modules/tracker/rbt/include/visp3/rbt/vpRBDenseDepthTracker.h:258:13: warning: conversion to ‘unsigned int’ from ‘int’ may change the sign of the result [-Wsign-conversion]
  258 |           L[i][0] = nX;
      |             ^
visp/modules/tracker/rbt/include/visp3/rbt/vpRBDenseDepthTracker.h:259:13: warning: conversion to ‘unsigned int’ from ‘int’ may change the sign of the result [-Wsign-conversion]
  259 |           L[i][1] = nY;
      |             ^
visp/modules/tracker/rbt/include/visp3/rbt/vpRBDenseDepthTracker.h:260:13: warning: conversion to ‘unsigned int’ from ‘int’ may change the sign of the result [-Wsign-conversion]
  260 |           L[i][2] = nZ;
      |             ^
visp/modules/tracker/rbt/include/visp3/rbt/vpRBDenseDepthTracker.h:261:13: warning: conversion to ‘unsigned int’ from ‘int’ may change the sign of the result [-Wsign-conversion]
  261 |           L[i][3] = nZ * Y - nY * Z;
      |             ^
visp/modules/tracker/rbt/include/visp3/rbt/vpRBDenseDepthTracker.h:262:13: warning: conversion to ‘unsigned int’ from ‘int’ may change the sign of the result [-Wsign-conversion]
  262 |           L[i][4] = nX * Z - nZ * X;
      |             ^
visp/modules/tracker/rbt/include/visp3/rbt/vpRBDenseDepthTracker.h:263:13: warning: conversion to ‘unsigned int’ from ‘int’ may change the sign of the result [-Wsign-conversion]
  263 |           L[i][5] = nY * X - nX * Y;
      |             ^
visp/modules/tracker/rbt/src/features/vpRBDenseDepthTracker.cpp: In member function ‘virtual void vpRBDenseDepthTracker::extractFeatures(const vpRBFeatureTrackerInput&, const vpRBFeatureTrackerInput&, const vpHomogeneousMatrix&)’:
visp/modules/tracker/rbt/src/features/vpRBDenseDepthTracker.cpp:80:52: warning: conversion to ‘unsigned int’ from ‘int’ may change the sign of the result [-Wsign-conversion]
   80 |       unsigned int numThreads = omp_get_num_threads();
      |                                 ~~~~~~~~~~~~~~~~~~~^~
visp/modules/tracker/rbt/src/features/vpRBDenseDepthTracker.cpp:90:48: warning: conversion to ‘unsigned int’ from ‘int’ may change the sign of the result [-Wsign-conversion]
   90 |     unsigned int threadIdx = omp_get_thread_num();
      |                              ~~~~~~~~~~~~~~~~~~^~
visp/modules/tracker/rbt/src/features/vpRBDenseDepthTracker.cpp:98:71: warning: conversion to ‘std::vector<vpRBDenseDepthTracker::vpDepthPoint>::size_type’ {aka ‘long unsigned int’} from ‘int’ may change the sign of the result [-Wsign-conversion]
   98 |     localPoints.reserve(m_depthPoints.capacity() / omp_get_num_threads());
      |                                                    ~~~~~~~~~~~~~~~~~~~^~
visp/modules/tracker/rbt/src/features/vpRBDenseDepthTracker.cpp:107:93: warning: conversion to ‘unsigned int’ from ‘int’ may change the sign of the result [-Wsign-conversion]
  107 |     for (auto i = static_cast<int>(bb.getTop()); i < static_cast<int>(bb.getBottom()); i += m_step) {
      |                                                                                             ^~~~~~
visp/modules/tracker/rbt/src/features/vpRBDenseDepthTracker.cpp:107:90: warning: conversion to ‘int’ from ‘unsigned int’ may change the sign of the result [-Wsign-conversion]
  107 |     for (auto i = static_cast<int>(bb.getTop()); i < static_cast<int>(bb.getBottom()); i += m_step) {
      |                                                                                        ~~^~~~~~~~~
visp/modules/tracker/rbt/src/features/vpRBDenseDepthTracker.cpp:108:95: warning: conversion to ‘unsigned int’ from ‘int’ may change the sign of the result [-Wsign-conversion]
  108 |       for (auto j = static_cast<int>(bb.getLeft()); j < static_cast<int>(bb.getRight()); j += m_step) {
      |                                                                                               ^~~~~~
visp/modules/tracker/rbt/src/features/vpRBDenseDepthTracker.cpp:108:92: warning: conversion to ‘int’ from ‘unsigned int’ may change the sign of the result [-Wsign-conversion]
  108 |       for (auto j = static_cast<int>(bb.getLeft()); j < static_cast<int>(bb.getRight()); j += m_step) {
      |                                                                                          ~~^~~~~~~~~

+++ modules/tracker/rbt/include/visp3/rbt/vpRBDenseDepthTracker.h
@@ -1,6 +1,6 @@
 /*
  * ViSP, open source Visual Servoing Platform software.
- * Copyright (C) 2005 - 2025 by Inria. All rights reserved.
+ * Copyright (C) 2005 - 2026 by Inria. All rights reserved.
  *
  * This software is free software; you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -86,7 +86,7 @@ public:
     DT_INVALID = 4
   };
 
-  vpRBDenseDepthTracker() : vpRBFeatureTracker(), m_step(2), m_maxFeatures(0), m_useMask(false), m_minMaskConfidence(0.f), m_displayType(vpDisplayType::DT_SIMPLE) { }
+  vpRBDenseDepthTracker() : vpRBFeatureTracker(), m_step(2), m_maxFeatures(0), m_useMask(false), m_minMaskConfidence(0.f), m_displayType(vpDisplayType::DT_SIMPLE) {}
 
   virtual ~vpRBDenseDepthTracker() = default;
 
@@ -152,19 +152,19 @@ public:
   /**
    * @brief Method called when starting a tracking iteration
    */
-  void onTrackingIterStart(const vpRBFeatureTrackerInput &, const vpHomogeneousMatrix & /*cMo*/) VP_OVERRIDE { }
-  void onTrackingIterEnd(const vpHomogeneousMatrix & /*cMo*/) VP_OVERRIDE { }
+  void onTrackingIterStart(const vpRBFeatureTrackerInput &, const vpHomogeneousMatrix & /*cMo*/) VP_OVERRIDE {}
+  void onTrackingIterEnd(const vpHomogeneousMatrix & /*cMo*/) VP_OVERRIDE {}
 
   void extractFeatures(const vpRBFeatureTrackerInput &frame, const vpRBFeatureTrackerInput &previousFrame, const vpHomogeneousMatrix &cMo) VP_OVERRIDE;
-  void trackFeatures(const vpRBFeatureTrackerInput &/*frame*/, const vpRBFeatureTrackerInput &/*previousFrame*/, const vpHomogeneousMatrix &/*cMo*/) VP_OVERRIDE { }
-  void initVVS(const vpRBFeatureTrackerInput &/*frame*/, const vpRBFeatureTrackerInput &/*previousFrame*/, const vpHomogeneousMatrix &/*cMo*/) VP_OVERRIDE { }
+  void trackFeatures(const vpRBFeatureTrackerInput &/*frame*/, const vpRBFeatureTrackerInput &/*previousFrame*/, const vpHomogeneousMatrix &/*cMo*/) VP_OVERRIDE {}
+  void initVVS(const vpRBFeatureTrackerInput &/*frame*/, const vpRBFeatureTrackerInput &/*previousFrame*/, const vpHomogeneousMatrix &/*cMo*/) VP_OVERRIDE {}
   void computeVVSIter(const vpRBFeatureTrackerInput &frame, const vpHomogeneousMatrix &cMo, unsigned int iteration) VP_OVERRIDE;
 
   void display(const vpCameraParameters &cam, const vpImage<unsigned char> &I, const vpImage<vpRGBa> &IRGB, const vpImage<unsigned char> &depth) const VP_OVERRIDE;
 
   struct vpDepthPoint
   {
-    vpDepthPoint() { }
+    vpDepthPoint() {}
   public:
     std::array<double, 3> oX, observation, objectNormal;
     std::array<double, 2> pixelPos;
@@ -210,7 +210,7 @@ public:
 #ifdef VISP_HAVE_OPENMP
 #pragma omp parallel for
 #endif
-      for (int i = 0; i < static_cast<int>(numPoints); ++i) {
+      for (unsigned int i = 0; i < numPoints; ++i) {
 
         //Step 1: update and filter out points that are no longer valid
         {
